requirements.txt
aiogram>=3.4,<4.0
aiosqlite>=0.19
python-dotenv>=1.0
google-generativeai>=0.7
requests>=2.31

